Keep demo lifecycle boundaries small and deterministic
Runtime command construction and supervisor execution are split into single-purpose helpers, while shared process fixtures remove duplicated setup and parallel test collisions. Cleanup signals no longer leak expected process-exit noise.\n\nConstraint: Repository functions and files must remain within the 25-line and 200-line limits\nRejected: Relax the hung-agent timing assertion | would hide fixture contention instead of controlling it\nConfidence: high\nScope-risk: narrow\nDirective: Keep process lifecycle tests serialized while they share host-level timing resources\nTested: All #7102 contracts; repeated supervisor contracts; strict kamn-e2e-harness clippy\nNot-tested: Full workspace gate and live Pi/devnet path pending
